promote string_to_filename to fontdrasil and use it instead of safe_filename
