+++ fontc/src/lib.rs
@@ -493,7 +493,7 @@ mod tests {
     use fontbe::orchestration::{
         AnyWorkId, Context as BeContext, Glyph, LocaFormatWrapper, WorkId as BeWorkIdentifier,
     };
-    use fontdrasil::{coords::NormalizedCoord, paths::safe_filename, types::GlyphName};
+    use fontdrasil::{coords::NormalizedCoord, paths::string_to_filename, types::GlyphName};
     use fontir::{
         ir::{self, GlyphOrder, KernGroup, KernPair, KernSide},
         orchestration::{Context as FeContext, Persistable, WorkId as FeWorkIdentifier},
@@ -1030,12 +1030,20 @@ mod tests {
     }
 
     fn read_ir_glyph(build_dir: &Path, name: &str) -> ir::Glyph {
-        let raw_glyph = read_file(&build_dir.join("glyph_ir").join(safe_filename(name, ".yml")));
+        let raw_glyph = read_file(
+            &build_dir
+                .join("glyph_ir")
+                .join(string_to_filename(name, ".yml")),
+        );
         ir::Glyph::read(&mut raw_glyph.as_slice())
     }
 
     fn read_be_glyph(build_dir: &Path, name: &str) -> RawGlyph {
-        let raw_glyph = read_file(&build_dir.join("glyphs").join(safe_filename(name, ".glyf")));
+        let raw_glyph = read_file(
+            &build_dir
+                .join("glyphs")
+                .join(string_to_filename(name, ".glyf")),
+        );
         let read: &mut dyn Read = &mut raw_glyph.as_slice();
         Glyph::read(read).data
     }
